refactor(datastore): add stub error maps
